+++ fortnitepy/http.py
@@ -82,7 +82,6 @@ class HTTPClient:
         if self.__session:
             await self.__session.close()
 
-
     async def request(self, method, url, is_json=False, **kwargs):
         headers = {**kwargs.get('headers', {}), **self.headers}
 
@@ -209,6 +208,81 @@ class HTTPClient:
         return await self.get(
             'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/' \
             'statsv2/account/{0}{1}'.format(user_id, query_parameters),
+            self.client.auth.authorization
+        )
+
+    async def get_multiple_br_stats_v2(self, ids, stats, start_time=None, end_time=None):
+        params = {}
+        if start_time:
+            params['startTime'] = start_time
+        if end_time:
+            params['endTime'] = end_time
+
+        return await self.post(
+            'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/statsv2/query',
+            self.client.auth.authorization,
+            data={
+                'owners': ids,
+                'stats': stats
+            },
+            is_json=True,
+            params=params
+        )
+
+    async def get_br_leaderboard(self, stat, platform, mode, window, page=0, items_per_page=50):
+        return await self.post(
+            'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/' \
+            'leaderboards/type/global/stat/br_{0}_{1}_m0_{2}/window/{3}?ownertype=1'.format(
+                stat,
+                platform.value,
+                mode.value,
+                window.value
+            ),
+            self.client.auth.authorization,
+            is_json=True,
+            params={
+                'pageNumber': page,
+                'itemsPerPage': items_per_page
+            }
+        )
+
+    async def get_br_leaderboard_v2(self, stat):
+        return await self.get(
+            'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/' \
+            'statsv2/leaderboards/{0}'.format(stat),
+            self.client.auth.authorization
+        )
+
+    async def get_lightswitch_status(self, service_id=None):
+        return await self.get(
+            'https://lightswitch-public-service-prod06.ol.epicgames.com/lightswitch/api/' \
+            'service/bulk/status',
+            self.client.auth.authorization,
+            params={'serviceId': service_id} if service_id else None
+        )
+
+    async def get_store_catalog(self):
+        return await self.get(
+            'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/storefront/v2/catalog',
+            self.client.auth.authorization
+        )
+
+    async def get_fortnite_news(self):
+        return await self.get(
+            'https://fortnitecontent-website-prod07.ol.epicgames.com/content/api/pages/fortnite-game',
+            self.client.auth.authorization
+        )
+
+    async def get_fortnite_content(self):
+        data = await self.get(
+            'https://fortnitecontent-website-prod07.ol.epicgames.com/content/api/pages/fortnite-game',
+            self.client.auth.authorization
+        )
+        return json.loads(data)
+
+    async def get_fortnite_timeline(self):
+        return await self.get(
+            'https://fortnite-public-service-prod11.ol.epicgames.com/fortnite/api/calendar/v1/timeline',
             self.client.auth.authorization
         )